+++ src/scripts/SpellHandlers/MiscSpells.cpp
@@ -405,22 +405,22 @@ bool CrystalSpikes(uint32 i, Spell* pSpell)
 
     Unit* pCaster = pSpell->u_caster;
 
-    for (int i = 1; i < 6; ++i)
+    for (uint8 i = 1; i < 6; ++i)
     {
         pCaster->GetMapMgr()->GetInterface()->SpawnCreature(CN_CRYSTAL_SPIKE, pCaster->GetPositionX() + (3 * i) + RandomUInt(2), pCaster->GetPositionY() + (3 * i) + RandomUInt(2), pCaster->GetPositionZ(), pCaster->GetOrientation(), true, false, 0, 0);
     }
 
-    for (int i = 1; i < 6; ++i)
+    for (uint8 i = 1; i < 6; ++i)
     {
         pCaster->GetMapMgr()->GetInterface()->SpawnCreature(CN_CRYSTAL_SPIKE, pCaster->GetPositionX() - (3 * i) - RandomUInt(2), pCaster->GetPositionY() + (3 * i) + RandomUInt(2), pCaster->GetPositionZ(), pCaster->GetOrientation(), true, false, 0, 0);
     }
 
-    for (int i = 1; i < 6; ++i)
+    for (uint8 i = 1; i < 6; ++i)
     {
         pCaster->GetMapMgr()->GetInterface()->SpawnCreature(CN_CRYSTAL_SPIKE, pCaster->GetPositionX() + (3 * i) + RandomUInt(2), pCaster->GetPositionY() - (3 * i) - RandomUInt(2), pCaster->GetPositionZ(), pCaster->GetOrientation(), true, false, 0, 0);
     }
 
-    for (int i = 1; i < 6; ++i)
+    for (uint8 i = 1; i < 6; ++i)
     {
         pCaster->GetMapMgr()->GetInterface()->SpawnCreature(CN_CRYSTAL_SPIKE, pCaster->GetPositionX() - (3 * i) - RandomUInt(2), pCaster->GetPositionY() - (3 * i) - RandomUInt(2), pCaster->GetPositionZ(), pCaster->GetOrientation(), true, false, 0, 0);
     }
